AP_IOMCU: Use a higher threshhold fo failing IO health.
We fail ~0.6% of all writes (on a 300Hz, disarmed quadplane),
which leads to failing the same percentage of attempts to arm the
aircraft as the IO processor is considered unhealthy. Given what the
processor doe this is a terrifying message as a user. The failure rate
climbs very rapidly which allows this to be a safe change.